[#5719] Add new utility method to improve plural language fallback
Fixes a gap in how we handle localizing plurals that results from
different languages having additional plural rules than English
which would result in localization keys being displayed because
English lacks the proper fallback string.

This new method takes the count as well as a function for producing
the final localization key from the PR. It will check to ensure the
produced string exists in the current language file, and if not it
will re-determine the appropriate plural rule as if the language
were set to English and return that localization key instead.

All places in the system where `getPluralRules` was used or a new
`Intl.PluralRules` object was constructed have been replaced with
calls to this new utility method (except for in consumption target
data, where some additional work will need to be done because of
how that code is structured).

Closes #5719