[#5164] Support other advancement types on features (#5289)
* [#5164] Support other advancement types on features

Adds enables `ItemChoice`, `ItemGrant`, `ScaleValue`, and `Trait`
advancement types to be added to `feat` items.

To better support grouping on the actor sheet, this also adds a
`ultimateOrigin` flag to items added via advancement that points
to the origin item in the advancement chain. So if a class adds
a feat which adds its own feature, then the `advancementOrigin` on
the feature will point to the feat, while the `ultimateOrigin` will
point to the class. A small exception to this logic has been added
for subclasses added directly by classes so that subclass features
point back to the subclass, not the class.

A point of future work is adding some control over the levels on
feat advancement. Currently the levels for those advancement types
refer to the character level, even if it is on a feature added by
a class. Potentially an option could be added to advancements on
feats that controls how their level information is determined.

Closes #5164

* [#5164] Rename to `advancementRoot`, use class/subclass level

Renames the `dnd5e.ultimateOrigin` flag to `dnd5e.advancementRoot`
for improved clarity.

Items that are granted by classes and subclasses and in turn have
their own advancement will now use the class level, rather than
the character level.

* [#5164] Opt feats out of class levels for advancement

Moves `advancementRootItem` into item system data model and adds
`advancementLevel` and `AdvancementClassLinked` to allow individual
item types to opt-out of linkage to class levels.

Did some adjustments to the step data and the step creation during
the advancement process to ensure the character level is available
when necessary.

* [#5164] Clean up minor code style issue

* [#5164] Fix some calls to `currentLevel`