[#5633] Improve calendar animations over large advances
Improves the calendar animation when advancing more than a single
day in time. In the previous animation the sun could be seen
moving the wrong direction across the sky if time advanced from
the afternoon of one day to the morning of the following day.

With this change the day & night progress variables are incremented
continuously, rather than being reset for each day. The sun
animation was tweaked to ensure it was in the correct position when
these ever increasing values were used. The formula for the sun
position adjustment is this:

`((1 - (x + .5 mod 2)) * 0.45) + x`

In order to properly support the increasing values several deltas
are added to the options during the `updateWorldTime` hook. These
values indicate how many times midnight, midday, sunrise, and
sunset have been passed during that time change (currently only
the midnight and midday deltas are used, but the sunrise and sunset
ones will be useful for item recharging).