[#5681] Fix error when using activity with enchanted consumption
The enchantment objects we are creating through our custom version
of `EmbeddedDataField` don't have a parent set because that data
isn't available in `_castChangeDelta`. This performs an additional
step to add the parent in to fix the error in consumption.

This shouldn't be necessary in the future when this is fixed by
https://github.com/foundryvtt/foundryvtt/issues/12528 in core so
long as core properly sets the parent.

Closes #5681