[#5174] Merge roll options more intelligently
Adds `mergeOptions` that is now called by `BasicRoll#mergeConfigs`
that allows specific roll types to customize the options merging
behavior. This allows D20 rolls to handle merging advantage and
range options in a logical manner. To ensure the proper method is
called, most calls to `BasicRoll#mergeConfigs` have been updated
to `D20Roll#mergeConfigs`.

This also removes the manual merging in `rollConcentration` that
was needed to handle those options.