[#5825] Add Epic Boon support to ability score advancement
Adds support for using the `AbilityScoreImprovementAdvancement` to
give an Epic Boon. When an ASI advancement is placed at level 19
or higher when the modern rules are used for the item it will be
automatically renamed "Epic Boon" and a recommendation field will
become available in the ASI config. This recommendation will be
shown as an additional option in the advancement flow interface
and displayed beneath the advancement in the item list.

The class summary journal page has also been updated to properly
list this as an Epic Boon and display the correct description in
the features list.

 ## Todo
- [ ] Update SRD classes to use ASI for epic boon

Closes #5825