[#5728] Fix sorting of spells when using "Priority" sort mode
Previously sorting was subtracting `preparation.mode` in the sort
method, but since that is a string it would result in `NaN`. This
uses a simple string comparison instead to resolve this issue.

This also changes the order for checking the `prepared` boolean so
prepared spells are sorted above non-prepared spells.

Closes #5728