[#5779] Add system-specific movement cost functions (#5780)
* [#5779] Add system-specific movement cost functions

Replaces the default movement cost functions provided by core with
versions that take the actor's movement types into account. If the
actor type has movement of that type then it will use a normal
cost. If the actor doesn't have that type but its walk speed can
be used in place of it, then it will increase the cost by 1 foot
for each foot. If walking speed isn't a valid fallback, then they
will not be permitted to use that movement type.

This includes an expansion of the `CONFIG.DND5E.movementTypes`
object to include the walk fallback data.

Closes #5779

* [#5779] Adjust some naming & register based on movement type config

* [#5779] Adjust animation speed to match movement cost