[#5171, #5173, #5174] Add advantage to abilities, skills & tools (#5271)
* [#5171, #5173, #5174] Add advantage to abilities, skills & tools

Adds the ability to configure roll mode, minimum, and maximum to
ability checks & saves, and roll mode for skills & tools.

Adds a new method `AdvantageModeField#combineFields` that takes
the key paths for two or more fields on a model and combines their
advantage counts to find a final advantage mode.

This new method is then used when rolling skill checks, tool
checks, or concentration saves to find the final advantage mode.

Closes #5171
Closes #5173

* Move roll mode below min/max because it looks nicer.

* [#5174] Merge roll options more intelligently

Adds `mergeOptions` that is now called by `BasicRoll#mergeConfigs`
that allows specific roll types to customize the options merging
behavior. This allows D20 rolls to handle merging advantage and
range options in a logical manner. To ensure the proper method is
called, most calls to `BasicRoll#mergeConfigs` have been updated
to `D20Roll#mergeConfigs`.

This also removes the manual merging in `rollConcentration` that
was needed to handle those options.

* [#5171] Apply ability check roll mode to initiative

---------

Co-authored-by: Kim Mantas <kim.mantas@gmail.com>