[#5300] Add feature prerequisites on other items (#5811)
Adds a list of identifiers to feature prerequisites and require
at least one of those other items to be present on the actor to
take that feature using an Choose Items or ASI advancement.

To support this functionality `identifiedItems` has been added
to actors along side `sourcedItems`. This groups items by their
identifiers and optionally allows getting only items with a
matching identifier and of a certain type.

Moves the prerequisite validation code into a new method on the
feature item itself to avoid code duplication between the two
advancement types.

Closes #5300