[#5825] Change SRD classes for Epic Boon ASIs, fix Primal Champion
Swaps the Choose Item advancement for Epic Boons with a Ability
Score Improvement advancement that includes the recommendation.
Remove the Grant Item advancement that gave the Epic Boon item for
each class (but left the old item around to avoid breaking classes
alreadyy on actors).

Also moved the Primal Champion ASI from the Barbarian class to the
Primal Champion item to avoid having it appear in the Barbarian
class journal page and prevent if from offerring a feat.