[#2187] Allow recovery periods to be set by uses scale value
Adds a scale value option to the recovery periods on items or
activities. For non-embedded items, selected the "Scale Value"
option will swap the input to a freeform text input with `@scale`
filled in, allowing for entering any scale value. For items on
actors, the list will be filled with any scale values on the actor
(unless the current value isn't found, in which case it will fall
back to showing the text input).

Closes #2187