[#5411] Allow transformation to retain spells in certain spell list
Adds a new transformation option that allows specifying one or more
spell lists. If included and the player has an item matching the
specified spell list identifier (so `subclass:moon` would search
for a subclass with the `moon` identifier), then any spells on that
list will be retained during transformation.

Adds spell lists to the SRD for each of the included subclasses to
allow for them to be selected.

Closes #5411