[#5137] Add spell list restrictions to Mystic Arcanum, tweak format
Changes the key used by spell list options to be separated by a
colon rather than a period to match the format used by traits.

Adjusted Mystic Arcanum advancement in SRD 5.1 to use a spell list
restriction and add advancement to the Mystic Arcanum item on the
SRD 5.2 to allow selection.