[#5787] Automatically create advancement on classes & backgrounds (#5866)
* [#5787] Automatically create advancement on classes & backgrounds

In addition to the default advancement created on Species, this
adds Trait & Grant Item advancement to backgrounds (and ASI with
the modern rules) and Hit Points, ASI, and a single Grant Items
advancement for classes.

Creates a new `AdvancementTemplate` for item data models that
encapsulates the default creation code so that individual types
just need to override `_advancementToCreate` to indicate what
should be created and call `preCreateAdvancement` during the
`_preCreate` method.

Closes #5787

* [#5787] Add subclass to class, remove extra languages