[#5581] Add direct link mode for selection in Transform activity (#5876)
* [#5581] Add direct link mode for selection in Transform activity

Adds mode to the Transform activity to select between the default
CR mode and a new Direct Link mode (similar to how the Summon
activity is set up). When the direct link mode is selected a UUID
field replaces the CR field allowing adding a specific actor which
will be used when transforming.

Closes #5581

* [#5581] Replace UUID input with custom input to match Summon UI

* [#5581] Adjust JSDoc for transform mode