[#3076] Allow all messages to pop-out, fix message title styling
Adds the `core.canPopout` flag to any chat message that doesn't
have it explicitly set.

Hides the title on popped-out chat messagges when they are expanded
and adjusts the styling when they are minimized. Also adds titles
to usage & item messages so that they will have a title when
minimized.

Closes #3076