[#4895] Add automatic self enchanting option for Enchant activity (#5841)
* [#4895] Add automatic self enchanting option for Enchant activity

Adds an option to the Behavior section of Enchant activities that
allows them to automatically enchant their parent item when used.
When activated with this option, the Enchant activity will check
to see if an enchantment from this item is already on its item. If
one is found, it will be removed. If one isn't found, or the
profile of the found enchantment doesn't match the configured
profile, then a new enchantment will be created. This allows for
easy toggling of an enchantment on an item, or swapping between
multiple enchantments on a single item.

When multiple profiles are available, the usage dialog will select
the currently applied one by default and indicate that it is active
in the list.

To avoid code duplication, some of the enchantment application code
has been moved from `EnchantmentApplicationElement` into the
`EnchantActivity`.

There is also some changes to how the behavior section of the
activity sheets are created to avoid having to create a new
template every time a new behavior option is required.

Closes #4895

* [#4895] Fix JSDoc on `applyEnchantment` method