[#5872] Adjust how tool bonuses are selected by check activity
Allows the check activity to make use of a tool's bonus so long as
the tool's base item matches the tool of the check being made, or
for any tool check if the item doesn't have a base item set.

Also removes the deprecation warning from `CONFIG.DND5E.toolIds`
because that method is still used by the system for displaying
base items on the item sheet, and it would be much more complicated
to rewrite the base item code rather than leave the proxy in place.

Closes #5872