[#5873] Move attunement counting into item's final data preparation
Since the previous method occured during the base data preparation
for actors any attunement added by enchantments wasn't available so
those items wouldn't be included in the count. This moves the
attunement counting to the items during their `prepareFinalData`
step. This also helps to remove duplicate code when checking if an
item can be attuned.

Closes #5873