[#5025] Switch legendaries to use `spent` rather than `value`
Change legendary actions and resistances from storing a `value` to
storing the number of `spent`. This leads to the correct number of
available legendaries when an NPC enters or leaves their lair.

Some adjustments were made to `consumeAttribute` to allow for
adjusting `spent` when the consumption target is `value`. This
helps ensure the consumption limits are correct and that no
migration is needed for consumption targets.

Closes #5025