[#5887] Fix using upcast spell scrolls from local spells
When using a spell scroll created from a non-compendium spell at
a higher level the `_prepareUsageScaling` method on activities
would attempt to set `flags.dnd5e.scaling` to handle the upcasting
but that flag was already at the correct value. Because the flag
wasn't updated, the item wasn't reset and so the subsequent call
to `prepareFinalAttributes` would be called on an item a second
time leading to an error.

Closes #5887