[#5411] Allow transformation to retain spells in certain spell list (#5878)
* [#5411] Allow transformation to retain spells in certain spell list

Adds a new transformation option that allows specifying one or more
spell lists. If included and the player has an item matching the
specified spell list identifier (so `subclass:moon` would search
for a subclass with the `moon` identifier), then any spells on that
list will be retained during transformation.

Adds spell lists to the SRD for each of the included subclasses to
allow for them to be selected.

Closes #5411

* Do not keep spells when Wild Shaped. Hide subclass lists from the compendium browser.

* [#5411] Add Keep Spells back to Wild Shape preset

* [#5411] Remove Keep Spells, use spell list length for copying spell slots

---------

Co-authored-by: Kim Mantas <kim.mantas@gmail.com>