[#5874] Add antimagic region behavior, tie into `magicAvailable`
Creates a new `antimagic` region behavior. When a token enters this
region it gains the new `antimagic` status effect. The effect is
then checked for in the `magicAvailable` method, disabling anything
that requires magic. It also hides any spells in the spells tab.

The `magicAvailable` method is moved from `EquippableItemTemplate`
to `ItemDataModel` so it is available on all items and the logic
has been updated to ensure it respects the new status effect and
works regardless of whether the item has properties or attunement.

Closes #5874