[#5994, #6013] Fix advancement for items added by subclasses
Ensures the proper class item is selected to determine proper
advacement levels for items added by subclasses. Originally
`advancementRootItem` for items added by subclasses was set to
the class, but `SubclassAdvancement` was changed to not set the
`rootItem` on the subclass to avoid different issues and support
dropping subclasses directly on the sheet, leading to this bug.

Closes #5994
Closes #6013