Use migration to set initial magical value, change default to false
Changes the default value of magical on effects to `false`, but use
the creation options when creating on a magical item to set it for
newly created effects.

Adds a migration to change effects that existed before this version
to have the magical effect if they are on spells, spell scrolls, or
an item with the `magical` property.