Add `ActiveEffectDataModel`, abstract some logic from document
Creates `ActiveEffectDataModel` that new can extend to gain some
default functionality. This new model includes getters for the
actor and item containing the effect, as well as `applicableType`
which returns the document type to which the effect should apply.
This allows some of the previous explicit checks for enchantments
and rider effects to be handled by the specific types and supports
module-provided AE types that can apply to actors, items, or
neither.

Also moves some of the enchantment-specific creation & deletion
logic into the system data model instead.