[#4772] Hide/suppress riders for disabled enchantments
When an enchantment that contains rider items or effects is
disabled, hide any rider items from the sheet, any rider activities
from the item, and suppress and rider effects.

Adds `enchantment.origin` flag to activities and effects to match
the flag added to rider items. This entailed adding support for
flags on activities.

Closes #4772