[#4772] Us IDs for rider activities/effects, fully hide effects
Use IDs rather than UUIDs for the flag on activities and effects
to make it resiliant if the item is moved between actors. Also
renamed the flag to `riderOrigin` to allow it to be used outside
of enchantments.

Modified the behavior of effects so they are fully hidden on the
sheet in addition to being suppressed.