[#6207, #6208] Respect damage threshold in application
Damage calculation will now respect damage thresholds found on the
actor. This adds the `ignore.threshold` boolean to damage options
so it can be ignored. To avoid having to re-sum the total damage
twice, damage summing has been moved into `calculateDamage` and
saved onto the result.

The damage application interface will display an icon for the
damage threshold if it is in effect so it can also be ignored
through the UI.

Adds damage threshold to character & NPC data (not currently
showing in UI, that will be enabled with a different PR).

Closes #6207
Closes #6208