[#5982] Roll encounter quantities when placing if currently `null`
When placing an encounter's actors in the scene, if any of the
quantities are set to `null` it will roll those before placing
them (but will not update the actor with these rolled values).

Since this doesn't require an actor update, this also makes the
"Place Members" button always interactive so encounters in locked
compendiums can still be used to populate the world.

Made a change to when the quantity is displayed on the sheet to
make the difference between a quantity of `0` and a `null` quantity
more distinct.

Also makes a change to the `rollQuantities` method so that the roll
button beside a single member doesn't have to duplicate the rolling
code.

Finally fixes a bug where clicking the "Place Members" buttons when
all members have a quantity of `0` would result in the sheet
minimizing and then immediately trying to maximize, throwing an
error.

Closes #5982