[#4962] Allow dynamic compendium browser filter choices
Allows the choices for a `set` filter in the compendium browser to
be a function rather than a static object. If a function is
provided, then it will be passed the currently set of filters so it
can be dynamically adjusted. This allows filtering the list of
feature types down to just the relevant ones, such as just the feat
types when viewing the "Feats" tab in standard mode.

Closes #4962