[#4957] Automatically scale down titles that don't fit
Adds some resizing of the title in all system applications to try
and prevent the title from being cut off. The title can be scaled
down up to 50%.

This is achieved by creating a second copy of the title that isn't
visible and isn't constrained by the window's width. On render or
resize the width of this invisible copy is compared against the
width of the window, and if it is wider then a scaling factor is
applied via CSS to the window's title.

Closes #4957