[#2291, #5770] Tie item uses recovery into calendar system
Allows recovery for for periods of "day", "dawn", or "dusk" to be
triggered by the passage of world time rather that through the
rest dialog. This is gated by a setting that by default is tied to
whether the calendar is enabled, but can be set to manual if
needed (for example, if using a module-provided calendar).

Some changes were made to the `recoverUses` API to support the
passage of multiple relevant recovery periods if the time was
advanced more than one day. This also fixes the bug with short
rests when using the gritty realism rest setting.

When a time passes the system checks to see if any relevant
recovery periods should trigger and if so, triggers recovery for
all world and actor items. If any changes were made, it then
creates a chat message to indicate the time passage and display
any deltas. These deltas are gated by observer permissions.

The created chat message also has a description of the time that
passed (e.g. "The sun has risen." or "Two days have passed."). In
order to support better localization of these times, some updates
have been made to the number formatting system to support the
`words` (and `numerals`) settings when formatting with units.

One potential point of improvement is how we handle selection of
recovery periods. At the moment, we only select the first recovery
profile that matches a provided period. This may cause issues for
cases where a "dawn" and "dusk" may have passed in one update. If
you had an item that was only usable during the daytime, you might
say it recovers all uses at dawn and loses them at dusk. This works
fine so long as you do not advance past both periods at the same
time. In the future the recovery uses method might be update with
the option of selecting multiple profiles so long as they are not
marked as mutually exclusive.

Closes #2291
Closes #5770