Do not diff item update in advancement manager to protect changes
Previously, all changes to advancement `value` data was applied
because the advancement were stored in an array which had to be
fully replaced with the update. Since changing it to an object
certain operations such as removing a key in Hit Points value data
wasn't being properly saved. This fixes that by forcing all changes
to be applied to items.