[#4708] Rebuild `AbilityScoreImprovementFlow` for V2
Rework `AbilityScoreImprovementFlow` for application V2 while
retaining the existing design.

The value data for ASIs has changed so `value.type` no longer
defaults to `asi`, it is instead set on the initial application if
relevant. This PR updates all of the ASIs in the SRD to get rid of
that initial value, but that isn't a required migration.