Disambiguate 2024 Channel Divinity identifiers, swap consumption
Changes the identifiers used for the 2024 Channel Divinity features
to make them different between the Cleric and the Paladin. This
ensures that any consumption tied to their identifiers will consume
the correct feature. This change wasn't made in the 2014 versions
because the multiclassing rules only allow you to use a single pool
of Channel Divinity.

Also updates consumption for features that consume both Channel
Divinity and Bardic Inspiration in both versions of the SRD to use
identifiers rather than UUIDs.