[#6264] Allow `type:identifier` format in feature prerequisites
Adds the ability to an `IdentifierField` to accept an item type
before the `:`. Only the identifier after the colon will be fully
validated.

The `identifiedItems#get` method will now check to see if a colon
is included in the provided key and split them up automatically to
make using it directly easier.

Makes the same adjustment to `SpellListRegistry#forType` to avoid
having to manually split identifiers in several places when dealing
with the compendium browser.

Closes #6264