Fix issues with negative consumption preview when using dice
Applies a few changes to how the preview for activity recovery
that includes die rolls. Previously if the first term was negative
and a die then the formula would display `(1d4 * -1) - 1` and it
would not scale properly using automatic mode. This adds `0 +` to
the start of formulas before scaling them to ensure the dice are
not multiplied by `-1`.

This also includes a new `BasicRoll#invert` method that takes the
roll's formula and inverts the operators so that the total from
the new formula will be the old total times negative one. This
allows the system to properly display non-deterministic recovery
formulas.