Add `adv` and `dis` die modifiers for advantage handling
Adds a new pair of die modifiers: `adv` for rolling with
advantage, and `dis` for rolling with disadvantage. These modifiers
simplify the normal setup so giving a roll advantage no longer
requires changing the number of dice rolled.

The modifier also supports a number, which indicates the number of
additional times the dice should be rolled, so `adv2` can be used
to handle Elven Accuracy, indicating that the original die formula
should be rolled 2 additional times and only the highest should be
taken.

The modifier also works for larger die pools, so if you wanted to
roll your damage with advantage you would just write `4d6adv` and
it would roll the dice twice, and take the set of four rolled that
has the highest total (different than changing it to `8d6kh4`,
which would select the highest four dice from all of those rolled
and would result in a different total).