[#6311] Convert dependent document tracking to registry
Switches to using a `dnd5e.dependentOn` flag on dependent documents
and a `dnd5e.registry.dependents` registry to track and handle
removing dependent documents, rather than a flag on the active
effect to which they are dependent.

For dependent activities and effects created directly on the item
the `dependentOn` flag is stored as a relative UUID to ensure that
duplicating an item with an enchantment will no lead to riders from
that enchantment being linked to the incorrect document.

Closes #6311