[#3748] Summon tokens with same elevation of summoning token
The `TokenPlacement` system not accepts an `origin` token in its
configuration. If provided, the elevation of the summoned tokens
will match the elevation of the origin.

In order to get the proper token when summoning, Activity gets the
new `getUsageToken` method which returns a single `TokenDocument`
in the current scene to be considered the token using the activity.
This method makes use of a slightly modified `getSceneTargets`
utility method that now optionally accepts an actor. If the actor
is provided, it will be used when finding tokens, prioritizing
selected tokens associated with that actor if present.

Closes #3748