[#6322] Add support for "All Damage" in resistances/etc.
Adds an "All Damage" option in resistances, immunities, and
vulnerabilties that supersedes any other damage type entered in
those configs. When the actor is prepared the system will clear
any other damage types in those values if "All Damage" is set.
It will also check and remove any damage types in reistances that
are also included in immunities to help with issues in the damage
application tray and to clean up the sheet.

Changes had to made to the damage calculation system to separate
the `active` state into type-specific activations VS. "All"
activations. This allows a new icon to be displayed in the tray
to control resistances/etc. for "All Damage". Because this change
made the `hasEffect` and `ignore` functions bigger and more
confusing, this PR moves them into separate private methods so
they can be better documented.

Also adds "All Damage" option in damage modification, but this is
applied in addition to any damage-specific modification. The
process of applying modifications has been changed to take into
account that a damage operation may have multiple entries with the
same type. Previously, the system would apply the same damage
modification to each separate entry of a damage type. It will now
keep a running total of the modification applied to avoid over
applying.

Closes #6322