[#6089] Fix Cast activity duplicating spells added by advancement
If a spell on an actor's sheet that was granted by an advancement
is added to a cast activity then it will be duplicated any time
the containing activity is modified. This was because the UUID
used within `sourcedItems` was set to the UUID of the item in the
compendium because it uses `flags.dnd5e.sourceId` set by the
advancement as first priority, but the cast activity was changing
`_stats.compendiumSource`.

To fix this issue, this modifies the order that the UUIDs are
fetched when preparing `sourcedItems`, prefering `compendiumSource`
first and only falling back to `sourceId` if that isn't found.
This also removes the older `core.sourceId` check since those
should be fully migrated by core by now.

Closes #6089