[#6311] Convert dependent document tracking to registry (#6312)
* [#6311] Convert dependent document tracking to registry

Switches to using a `dnd5e.dependentOn` flag on dependent documents
and a `dnd5e.registry.dependents` registry to track and handle
removing dependent documents, rather than a flag on the active
effect to which they are dependent.

For dependent activities and effects created directly on the item
the `dependentOn` flag is stored as a relative UUID to ensure that
duplicating an item with an enchantment will no lead to riders from
that enchantment being linked to the incorrect document.

Closes #6311

* [#6311] Protect from using flag to delete documents w/o permission

Restrict any dependents defined in the old `dnd5e.dependents` flag
from returning unless the AE on which the flag is defined and the
document share a common ancestor document or the document is an
effect with its origin pointing to the AE with the flag.

* [#6311] Fix some documentation issues, clean up code style

* [#6311] Swap to using normal ID for storing same-item dependent flags

* [#6311] Tweak variable naming in registry