[#4354, #4772] Hide/suppress riders for disabled enchantments (#6176)
* [#4772] Hide/suppress riders for disabled enchantments

When an enchantment that contains rider items or effects is
disabled, hide any rider items from the sheet, any rider activities
from the item, and suppress and rider effects.

Adds `enchantment.origin` flag to activities and effects to match
the flag added to rider items. This entailed adding support for
flags on activities.

Closes #4772

* [#4772] Us IDs for rider activities/effects, fully hide effects

Use IDs rather than UUIDs for the flag on activities and effects
to make it resiliant if the item is moved between actors. Also
renamed the flag to `riderOrigin` to allow it to be used outside
of enchantments.

Modified the behavior of effects so they are fully hidden on the
sheet in addition to being suppressed.

* [#6198] Ensure combat activations aren't displayed if disabled

- Combat activations won't be listed if an activity isn't available
  to be used.
- Recovery of items and activities isn't performed if the rider
  origin is disabled

Closes #6198

* [#4772, #6198] Swap to using `dependentOn` data rather than flag

Uses `flags.dnd5e.dependentOn` rather than creating a new flag and
renames `riderOrigin` to `dependentOrigin` for consistency. Also
switches to checking `active === false` rather than `disabled` to
allow for future suppressed enchantment effects.