[#4963] Add visibility options to activities (#5892)
* [#4963] Add visibility options to activities

Adds several options to activities that control the ability to use
or configure them:
- Requires Attunement (not usable if not attuned)
- Requires Magic (not usable in antimagic field)
- Requires Identification (not usable or visible if not identified)
- Level range (not usable if player is outside the specified range,
  can be set to character level or specific class level)

When an activity isn't usable it will not appear in the activity
selection dialog, in the list of activities beneath an item in
inventory, or in the item's summary in inventory.

When an activity isn't configurable it won't appear in the list of
activities on the item (currently the activities tab is hidden on
unidentified items, but this might allow us to show it).

Some options aren't always available:
- Require attunement is only available if the item can be attuned,
  and will be locked to the same value as requires magic if the
  item requires attunement because by the rules, if attunement is
  required then all magical features of the item are locked behind
  attunement
- Requires magic is only available if the item has the "magical"
  property, unless magical isn't in the list of valid properties
- Requires identification is only available for items that can
  be identified
- Cast activity always has "Requires Magic" checked

Since `TransformActivity` and `SummonActivity` already had level
limits, the associated identifier has been migrated to match the
new visibility identifier available on all activities (the level
limits are still available on each profile, since those are still
useful).

Closes #4963

* Ensure all activity visibility options appear for rider activities

Doesn't modify the visibility values on rider activities and always
displays the visibility controls regardless of their parent's
details.

* [#4963] Adjust localization, require magic/attunement on cast

- Adjusted the class identifier hint
- Ensure the "Require Magic" field on cast activities is checked
- Check the "Require Attunement" field on cast activities if it
  is required
- Code style changes