[#4570] Allow consumed hit dice to be used in roll formulas (#6294)
When clicking on a chat card action or triggering subsequent
actions a flag will be set on the item that contains a formatted
version of the activity's consumption. Currently this only includes
a formula for the hit dice consumed at `@consumed.hd`, allowing any
hit dice consumed to be rolled.

Closes #4570