[#6376] Redesign Roll Table sheet using dnd5e2 styling
Implements a system version of the roll table sheet that uses the
system's styling. This doesn't both rewriting the HTML, but makes
certain targeted changes on render to ensure the proper classes
are added and the dnd5e checkboxes are used.

Core's edit/view toggle is replaced with an edit slider in the
header like other 5e sheets. In order to make this easier, some of
the system's mode toggle code was changed to be more similar to
how core's sheet is implemented (adding `isEditMode` property and
using a `changeMode` action rather than a direct listener. None of
these should be breaking changes.

Some changes were also made to sheet styles to make them more
widely available and item's `.item-image` header styling was made
into a more generic `.document-image` to be usable across sheets.

Closes #6376