[#6335] Add `ModifyItem` advancement for enchanting existing items
Adds a new advancement type that allows for applying an enchantment
to an existing item on an actor's sheet. This advancement contains
a list of enchantments to apply, and a list of identifiers that are
used to find the items to apply those enchantments to.

During the advancement process, each enchantment will be displayed
with a list of matching items below it.

 ### Future Improvement
- Allow user control over what items are actually enchanted
- Support rider activities & effects once #6357 is resolved

Closes #6335