[#5388] Allow "Lose All" recovery & "Destroy on Empty" to interact (#6214)
Adds handling to uses recovery to allow a "Lose All" recovery mode
and consumable's "Destroy on Empty" to interact, removing any items
that lose all charges through that recovery mode. This works for
both resting and combat-based recovery.

In order to properly communicate the item deletion this adds
support for created & deleted items to `ActorDeltasField`, which
also allows the `getDeltas` method to be used for activity usage,
removing some duplicate deltas code.

Closes #5388