[#3077] Swap chat icon based on preferred sheet artwork (#6178)
If a character or NPC has their token displayed on the sheet rather
than their portrait, display that in the chat card also. Moves the
logic for selecting the artwork into its own method so it can be
shared between the sheet and the chat card, and cache the result so
each chat card doesn't end up with a different random token.

Closes #3077