[#5633] Add standard calendar HUD (#5634)
* [#5633] Add standard calendar HUD

Introduces a new calendar HUD UI element that displays at the top
of the screen and is styled like the rest of the core UI. This
element includes buttons for advancing and reversing time, opening
the player's assigned character sheet, and opening the party group
sheet. Between that it displays the current date and time along
with a day visualizer to indicate the progress of the day.

The calendar interface is set by a `CONFIG.DND5E.calendar` object
that defines the application used to render the calendar, the
current HUD instance if rendered, a list of formatters that can be
used to display the date and time, and a pair of methods for
tracking the progress through the day & night. This makes it easy
to completely disable the native HUD (setting `application` to
`null`) or to customize it to whatever calendar is being used.

The settings offer two sets of calendar options, one visible only
to GM users and the other available to all users. The GM settings
allow for enabling the calendar for all users and for customizing
the time change provided by each of the advance/reverse buttons.

The preferences have a control to toggle the display of the HUD
for each user and options to customize what formatter is used to
display the date or time on each side of the widget.

 ## Todo
- [ ] Show/hide sheet buttons based on sheet changes
- [ ] Improve animation when advancing over large periods of time
- [ ] Give GMs more control over what formatters players can use
- [ ] Add UI for changing to specific date/time

* [#5634] Create `CalendarData5e` and move formatters & day progress

Creates a new `CalendarData5e` class to replace the default class.
This new class holds the new formatting methods and the progress
calculation methods. Some extra care was taken when calling the
progress methods in case the user has a `CalendarData` class that
doesn't provide them.

Also adjusts how the HUD insertion occurs to match core better.

* [#5633] Add specific sunrise/set methods to `CalendarData5e`

Adds `sunrise` and `sunset` methods that return the time for those
events for a given day and then adjusts the progress methods to
take advantage of those. These specific methods will be useful for
future features that trigger based on sunrise or sunset occurring.

* [#5633] Modify button preparation to be closer to header buttons

Modifies how the calendar buttons are prepared to be closer to how
the header buttons on V2 applications are handled:

- Moved `action` out of `dataset` to top level config
- Renamed `display` to `visible` and accept functions
- Added `onClick` to handle module-added buttons
- Used `ApplicationV2._doEvent` to trigger the button hook

* [#5633] Improve calendar animations over large advances

Improves the calendar animation when advancing more than a single
day in time. In the previous animation the sun could be seen
moving the wrong direction across the sky if time advanced from
the afternoon of one day to the morning of the following day.

With this change the day & night progress variables are incremented
continuously, rather than being reset for each day. The sun
animation was tweaked to ensure it was in the correct position when
these ever increasing values were used. The formula for the sun
position adjustment is this:

`((1 - (x + .5 mod 2)) * 0.45) + x`

In order to properly support the increasing values several deltas
are added to the options during the `updateWorldTime` hook. These
values indicate how many times midnight, midday, sunrise, and
sunset have been passed during that time change (currently only
the midnight and midday deltas are used, but the sunrise and sunset
ones will be useful for item recharging).

* [#5633] Add additional hover buttons for calendar time control

Adds a dropdown beneath the advance & reverse time buttons in the
calendar that give more time options. In the process this removes
the settings for configuring the time control buttons and reduces
it to a single visible time button on each side.

* [#5633] Don't fully capitalize calendar button text

* [#5633] Improve calendar buttons, change location, more

- Switch to using `game.actors.party` to get primary party
- Switch to using `_canRender` rather than modifying `_render`
- Adjust hints
- Set minimum width for calendar to prevent too much jumping
- Replace 30 minutes with 1 hour in time dropdown
- Allow default time button action to be customized
- Improve how buttons trigger actions
- Fix issue with advance time default action

* [#5633] Fix button layout issue, add 10 minute interval