Allow relative UUIDs in `dependentOn` flag on items
Opens up the possibility of having an item being dependent on an
enchantment that is contained on the item itself. A small change
has been made to the item hiding logic to ensure that items that
contain their own dependent origin are not fully hidden when that
dependent effect is disabled, otherwise it would be easy to hide
an item by disabling its enchantment and never be able to get it
back.