Add Eberron, Faerûn, and Greyhawk calendars, approximate formatters (#6415)
* Add Eberron, Faerûn, and Greyhawk calendars, approximate formatters

Adds three new calendars into the system for use in popular
settings: the Calendar of Greyhawk (for Greyhawk), the Calendar of
Harptos (for Forgotten Realms), and the Calendar of Khorvaire
(for Eberron).

The Calendar of Harptos includes some special formatting for the
six festval days that occur throughout the year outside of the
normal months. For those, it will just display the name of the
festival rather than the day and month.

This also includes two new formatters for approximate dates and
times. The date formatter uses the current season with "Early",
"Mid", and "Late" to give a hint as to when in the season it is.
For time it uses the time of sunrise and sunset and selects a
descriptive term depending on the time of the day. These terms
include Sunrise, Sunset, Noon, Midnight, Dawn, Dusk, Morning,
Afternoon, Evening, and Night.

* Fix issue with calendar data documentation

* Fix mis-renamed method call