Add dialog for setting the current date (#6433)
* Add dialog for setting the current date

Adds a new dialog for setting the game time to a specific date
using the year, month, and day. Because core's `GameTime#set`
method doesn't support using `month` or `dayOfMonth` to set the
date, this adds a new `CalendarData5e#jumpToDate` method that
takes those values and calculates the correct values before passing
them to core.

* Avoid mutating `game.time.components`