Add ability to enchant directly from compendium (#6449)
When an item from a compendium is dragged on the enchant tray it
will now be duplicated and a copy created on the actor that used
the enchant activity (if the player is the owner) or the player's
selected token or assigned character.

The item that is created in this manner is automatically made
dependent on its own effect, ensuring that if the enchantment is
disabled the created item will also be removed.