[#6528] Add flag for controlling NPC gear/features categorization
Adds a new flag `dnd5e.statBlockOverride.section` for individual
items that takes the options `gearOnly` and `featuresOnly`. If
`gearOnly` is set, then the item will appear in the gear list on
the NPC embeds, but not in the Actions section of the stat block
of on the Features tab of the NPC sheet. If `featuresOnly` is set
then the item will appear in the Actions section of the stat block
and the Features tab of the NPC sheet but not in the gear list.

Both of these settings will only cause the change to occur if the
item is already valid in that area, so setting this flag on a
feature item won't cause it to appear in the Gear list because it
isn't a physical item.

Closes #6528