Add application for configuring overall Table of Contents flags
Adds a new configuration application available from the header of
the table of contents window that displays all of the journal
entries in the pack grouped by folder and allows for setting the
`type` and `position` flag.

Configuration of the other entry flags (`order`, `title`, &
`showPages`) as well as the individual page flags (`tocHidden`)
will be left to a separate config application accessible from the
individual entry sheets.