[#6590] Fix cast activity not scaling when method not set on spell
Adds a spellcasting method override to the spell changes with
cast activity to ensure it always has the "Spellcasting" method
set. If no spellcasting method was set, then the spell wouldn't
scale properly (this only happened if the original spell had no
method set and it was added to an actor without spellcasting
classes).

Closes #6590